Move non UI code from ui to presentation package
