+++ tasks-app-shared/src/androidMain/kotlin/net/opatry/tasks/app/ui/component/CompletedTaskRowPreview.kt
@@ -30,8 +30,8 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.todayIn
-import net.opatry.tasks.app.ui.model.TaskId
-import net.opatry.tasks.app.ui.model.TaskUIModel
+import net.opatry.tasks.app.presentation.model.TaskId
+import net.opatry.tasks.app.presentation.model.TaskUIModel
 import net.opatry.tasks.app.ui.tooling.TaskfolioThemedPreview
 
 private class CompletedTaskRowPreviewDataProvider :
