+++ tasks-app-shared/src/androidMain/kotlin/net/opatry/tasks/app/ui/screen/tasksPanePreview.kt
@@ -38,8 +38,8 @@ import kotlinx.datetime.plus
 import kotlinx.datetime.todayIn
 import net.opatry.tasks.app.ui.component.CompletedTaskRow
 import net.opatry.tasks.app.ui.component.RemainingTaskRow
-import net.opatry.tasks.app.ui.model.TaskId
-import net.opatry.tasks.app.ui.model.TaskUIModel
+import net.opatry.tasks.app.presentation.model.TaskId
+import net.opatry.tasks.app.presentation.model.TaskUIModel
 import net.opatry.tasks.app.ui.tooling.TaskfolioThemedPreview
