+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/component/TaskListTopAppBar.kt
@@ -42,9 +42,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.style.TextOverflow
+import net.opatry.tasks.app.presentation.model.TaskListUIModel
 import net.opatry.tasks.app.ui.component.TaskListTopAppBarTestTag.MORE_MENU_ICON
 import net.opatry.tasks.app.ui.component.TaskListTopAppBarTestTag.SORT_MENU_ICON
-import net.opatry.tasks.app.ui.model.TaskListUIModel
 import net.opatry.tasks.data.TaskListSorting
 
 @VisibleForTesting
