+++ tasks-app-shared/src/commonTest/kotlin/net/opatry/tasks/ui/component/DateRangeColorTest.kt
@@ -30,7 +30,7 @@ import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus
 import net.opatry.tasks.app.ui.component.toColor
-import net.opatry.tasks.app.ui.model.DateRange
+import net.opatry.tasks.app.presentation.model.DateRange
 import net.opatry.tasks.ui.screen.today
 import kotlin.test.Test
 import kotlin.test.assertEquals
